refactor: split useCrossDomainRedirect into platform-specific files
Extract shared core logic into crossDomainRedirectUtils.ts and split
the hook into .ts (web/ext noop), .native.ts, and .desktop.ts to
eliminate runtime platformEnv checks and align with project conventions.
