+++ packages/kit/src/hooks/useCrossDomainRedirect.ts
@@ -1,107 +1,16 @@
-import { useCallback, useEffect, useMemo, useRef } from 'react';
-
-import { ipcMessageKeys } from '@onekeyhq/desktop/app/config';
-import platformEnv from '@onekeyhq/shared/src/platformEnv';
-import { openUrlInDiscovery } from '@onekeyhq/shared/src/utils/openUrlUtils';
-
-import useAppNavigation from './useAppNavigation';
+import type { ICrossDomainRedirectResult } from './crossDomainRedirectUtils';
 
 /**
- * Hook that intercepts cross-domain navigation in a WebView and redirects
- * them to the Discovery browser. Handles three interception mechanisms:
- *
- * - Desktop IPC: Electron main process intercepts window.open() and sends
- *   the URL via WEBVIEW_NEW_WINDOW IPC event
- * - Native onShouldStartLoadWithRequest: intercepts top-frame navigation
- * - Native onOpenWindow: intercepts window.open() popups
+ * Cross-domain redirect hook — web/extension default (noop).
+ * Web and extension platforms open fiat crypto URLs externally,
+ * so cross-domain interception is not applicable.
  */
-export function useCrossDomainRedirect(initialUrl: string, enabled = true) {
-  const navigation = useAppNavigation();
-  const isUnmounting = useRef(false);
-
-  useEffect(
-    () => () => {
-      isUnmounting.current = true;
-    },
-    [],
-  );
-
-  const initialHost = useMemo(() => {
-    try {
-      return new URL(initialUrl).hostname;
-    } catch {
-      return '';
-    }
-  }, [initialUrl]);
-
-  const isCrossDomain = useCallback(
-    (targetUrl: string) => {
-      try {
-        const targetHost = new URL(targetUrl).hostname;
-        return !!(targetHost && initialHost && targetHost !== initialHost);
-      } catch {
-        return false;
-      }
-    },
-    [initialHost],
-  );
-
-  const redirectToDiscovery = useCallback(
-    (targetUrl: string) => {
-      openUrlInDiscovery({ url: targetUrl });
-      navigation.pop();
-    },
-    [navigation],
-  );
-
-  // Desktop: intercept window.open() via Electron IPC
-  useEffect(() => {
-    if (!enabled || !platformEnv.isDesktop) return;
-    const handleDesktopNewWindow = (
-      _event: unknown,
-      data: { url?: string },
-    ) => {
-      if (isUnmounting.current || !data.url) return;
-      if (isCrossDomain(data.url)) {
-        redirectToDiscovery(data.url);
-      }
-    };
-    globalThis.desktopApi?.addIpcEventListener(
-      ipcMessageKeys.WEBVIEW_NEW_WINDOW,
-      handleDesktopNewWindow,
-    );
-    return () => {
-      globalThis.desktopApi?.removeIpcEventListener(
-        ipcMessageKeys.WEBVIEW_NEW_WINDOW,
-        handleDesktopNewWindow,
-      );
-    };
-  }, [enabled, isCrossDomain, redirectToDiscovery]);
-
-  // Native: intercept top-frame navigation
-  const onShouldStartLoadWithRequest = useCallback(
-    (event: { url: string; isTopFrame?: boolean }) => {
-      if (!enabled) return true;
-      if (!event.isTopFrame) return true;
-      if (isCrossDomain(event.url)) {
-        redirectToDiscovery(event.url);
-        return false;
-      }
-      return true;
-    },
-    [enabled, isCrossDomain, redirectToDiscovery],
-  );
-
-  // Native: intercept window.open() popups
-  const onOpenWindow = useCallback(
-    (event: { nativeEvent: { targetUrl: string } }) => {
-      if (!enabled) return;
-      if (isCrossDomain(event.nativeEvent.targetUrl)) {
-        redirectToDiscovery(event.nativeEvent.targetUrl);
-      }
-    },
-    [enabled, isCrossDomain, redirectToDiscovery],
-  );
-
-  return { onShouldStartLoadWithRequest, onOpenWindow };
+export function useCrossDomainRedirect(
+  _initialUrl: string,
+  _enabled = true,
+): ICrossDomainRedirectResult {
+  return {
+    onShouldStartLoadWithRequest: undefined,
+    onOpenWindow: undefined,
+  };
 }
